Sync upstream

+++ anomalib/models/components/dimensionality_reduction/random_projection.py
@@ -78,7 +78,8 @@ class SparseRandomProjection:
                 c_idx = torch.tensor(
                     sample_without_replacement(
                         n_population=n_features, n_samples=nnz_idx, random_state=self.random_state
-                    )
+                    ),
+                    dtype=torch.int64,
                 )
                 data = torch.distributions.Binomial(total_count=1, probs=0.5).sample(sample_shape=c_idx.size()) * 2 - 1
                 # assign data to only those columns
@@ -98,7 +99,7 @@ class SparseRandomProjection:
             eps (float, optional): Minimum distortion rate. Defaults to 0.1.
         """
 
-        denominator = (eps ** 2 / 2) - (eps ** 3 / 3)
+        denominator = (eps**2 / 2) - (eps**3 / 3)
         return (4 * np.log(n_samples) / denominator).astype(np.int64)
 
     def fit(self, embedding: Tensor) -> "SparseRandomProjection":
